+++ src/engine/engineVariables.ts
@@ -132,19 +132,7 @@ export class EngineVariables {
         opArr[name] = clientPackage;
     }
 
-    public getTestClientPackages(): { [id: string] : UniteClientPackage } {
-        const packages: { [id: string] : UniteClientPackage } = {};
-
-        Object.keys(this._requiredClientPackages)
-            .filter(key => this._requiredClientPackages[key].includeMode === "test" || this._requiredClientPackages[key].includeMode === "both")
-            .forEach(key => {
-                packages[key] = this._requiredClientPackages[key];
-            });
-
-        return packages;
-    }
-
-   public toggleDevDependency(dependencies: string[], required: boolean): void {
+    public toggleDevDependency(dependencies: string[], required: boolean): void {
         let opArr: string[];
         if (required) {
             opArr = this._requiredDevDependencies;