+++ src/pipelineSteps/taskManager/gulp.ts
@@ -203,6 +203,7 @@ export class Gulp extends EnginePipelineStepBase {
         this.toggleFile(assetUtils, "client-packages.js", this._utilFolder, "client-packages.js", isGulp);
         this.toggleFile(assetUtils, "display.js", this._utilFolder, "display.js", isGulp);
         this.toggleFile(assetUtils, "exec.js", this._utilFolder, "exec.js", isGulp);
+        this.toggleFile(assetUtils, "json-helper.js", this._utilFolder, "json-helper.js", isGulp);
         this.toggleFile(assetUtils, "package-config.js", this._utilFolder, "package-config.js", isGulp);
         this.toggleFile(assetUtils, "platform-utils.js", this._utilFolder, "platform-utils.js", isGulp);
         this.toggleFile(assetUtils, "theme-utils.js", this._utilFolder, "theme-utils.js", isGulp);