+++ src/pipelineSteps/unitTestEngine/chromeHeadless.ts
@@ -1,6 +1,7 @@
 /**
  * Pipeline step to generate chrome headless configuration.
  */
+import { ArrayHelper } from "unitejs-framework/dist/helpers/arrayHelper";
 import { IFileSystem } from "unitejs-framework/dist/interfaces/IFileSystem";
 import { ILogger } from "unitejs-framework/dist/interfaces/ILogger";
 import { KarmaConfiguration } from "../../configuration/models/karma/karmaConfiguration";
@@ -14,11 +15,10 @@ export class ChromeHeadless extends EnginePipelineStepBase {
                                             super.condition(uniteConfiguration.unitTestRunner, "Karma") &&
                                             super.condition(uniteConfiguration.unitTestEngine, "ChromeHeadless"));
 
-        if (super.condition(uniteConfiguration.unitTestRunner, "Karma") && super.condition(uniteConfiguration.unitTestEngine, "ChromeHeadless")) {
-            const karmaConfiguration = engineVariables.getConfiguration<KarmaConfiguration>("Karma");
-            if (karmaConfiguration) {
-                karmaConfiguration.browsers.push("ChromeHeadless");
-            }
+        const karmaConfiguration = engineVariables.getConfiguration<KarmaConfiguration>("Karma");
+        if (karmaConfiguration) {
+            ArrayHelper.addRemove(karmaConfiguration.browsers, "ChromeHeadless",
+                                  super.condition(uniteConfiguration.unitTestRunner, "Karma") && super.condition(uniteConfiguration.unitTestEngine, "ChromeHeadless"));
         }
 
         return 0;