+++ src/pipelineSteps/unitTestEngine/phantomJs.ts
@@ -1,6 +1,7 @@
 /**
  * Pipeline step to generate phantom JS configuration.
  */
+import { ArrayHelper } from "unitejs-framework/dist/helpers/arrayHelper";
 import { IFileSystem } from "unitejs-framework/dist/interfaces/IFileSystem";
 import { ILogger } from "unitejs-framework/dist/interfaces/ILogger";
 import { KarmaConfiguration } from "../../configuration/models/karma/karmaConfiguration";
@@ -13,15 +14,16 @@ export class PhantomJs extends EnginePipelineStepBase {
         engineVariables.toggleDevDependency(["karma-phantomjs-launcher", "bluebird"],
                                             super.condition(uniteConfiguration.unitTestRunner, "Karma") && super.condition(uniteConfiguration.unitTestEngine, "PhantomJS"));
 
-        if (super.condition(uniteConfiguration.unitTestRunner, "Karma") && super.condition(uniteConfiguration.unitTestEngine, "PhantomJS")) {
-            const karmaConfiguration = engineVariables.getConfiguration<KarmaConfiguration>("Karma");
-            if (karmaConfiguration) {
-                karmaConfiguration.browsers.push("PhantomJS");
+        const karmaConfiguration = engineVariables.getConfiguration<KarmaConfiguration>("Karma");
+        if (karmaConfiguration) {
+            const addCond = super.condition(uniteConfiguration.unitTestRunner, "Karma") && super.condition(uniteConfiguration.unitTestEngine, "PhantomJS");
 
-                const bbInclude = fileSystem.pathToWeb(
-                    fileSystem.pathFileRelative(engineVariables.wwwRootFolder, fileSystem.pathCombine(engineVariables.www.packageFolder, "bluebird/js/browser/bluebird.js")));
-                karmaConfiguration.files.push({ pattern: bbInclude, included: true });
-            }
+            ArrayHelper.addRemove(karmaConfiguration.browsers, "PhantomJS", addCond);
+
+            const bbInclude = fileSystem.pathToWeb(
+                fileSystem.pathFileRelative(engineVariables.wwwRootFolder, fileSystem.pathCombine(engineVariables.www.packageFolder, "bluebird/js/browser/bluebird.js")));
+
+            ArrayHelper.addRemove(karmaConfiguration.files, { pattern: bbInclude, included: true }, addCond, (obj, item) => obj.pattern === item.pattern);
         }
 
         return 0;