+++ test/unit/src/pipelineSteps/taskManager/gulp.spec.ts
@@ -159,7 +159,7 @@ describe("Gulp", () => {
 
             Chai.expect(res).to.be.equal(0);
 
-            Chai.expect(stub.callCount).to.be.equal(23);
+            Chai.expect(stub.callCount).to.be.equal(24);
 
             const packageJsonDevDependencies: { [id: string]: string } = {};
             engineVariablesStub.buildDevDependencies(packageJsonDevDependencies);
@@ -185,7 +185,7 @@ describe("Gulp", () => {
 
             Chai.expect(res).to.be.equal(0);
 
-            Chai.expect(stub.callCount).to.be.equal(27);
+            Chai.expect(stub.callCount).to.be.equal(28);
 
             const packageJsonDevDependencies: { [id: string]: string } = {};
             engineVariablesStub.buildDevDependencies(packageJsonDevDependencies);
@@ -211,7 +211,7 @@ describe("Gulp", () => {
 
             Chai.expect(res).to.be.equal(0);
 
-            Chai.expect(stub.callCount).to.be.equal(28);
+            Chai.expect(stub.callCount).to.be.equal(29);
 
             const packageJsonDevDependencies: { [id: string]: string } = {};
             engineVariablesStub.buildDevDependencies(packageJsonDevDependencies);