+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * HTML Template Script includes are generated at build time not configure
 * Remove unit-ui gulp task added --browser args to gulp unit
 * Added Firefox, Safari, IE launchers for Karma
+* Karma read current config and modify not overwrite
+* Karma generates include files at unit test time not configure
+* Removed karma pipeline step during clientConfiguration commands
 
 ## v0.9.6